+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 8.2.0 - 2025-04-10
+### Added
+- Functionality to allow renames in glue event listener. Table will be created with new name, copied partitions 
+from old to new table, and dropped old table.
+
 ## 8.1.2 - 2025-04-08
 ### Fix
 - Hive version validation when multiple listeners are used.